seperate class controller for popup attempt
